requirements.txt
Django>=4.2.0
channels>=4.0.0
channels-redis>=4.1.0
django-redis>=5.3.0
djangorestframework>=3.14.0
django-cors-headers>=4.1.0
python-dotenv>=1.0.0
textblob>=0.17.1
jieba>=0.42.1
gensim>=4.3.1
numpy>=1.24.3
scikit-learn>=1.2.2
torch>=2.0.1
transformers>=4.30.2
Pillow==10.3.0
pytest==8.1.1
pytest-django==4.8.0
openai==1.20.0
drf-yasg==1.21.7
jsonschema==4.21.1
daphne==4.0.0
python-json-logger==2.0.7 